feat(parser): Parse color values with meaning
Also added a bunch of access and modification methods

BREAKING CHANGE: ColorValues created with hex strings are now named as "hex" instead of "rgb", as we're now using tinycolor2 to parse the color values. If you only want a color value to display, use hexcolor() on the Color node instead.